создание страницы выбора сложности
