+++ index.html
@@ -9,5 +9,5 @@
 <body>
     <div id="app"></div>
 </body>
-    <script src="./index.js"></script>
+    <script type="module" src="./index.js"></script>
 </html>
